Adding Class=Active, Basename Function & Super Global Variable
Adding the Class=“active”:  We use a class called active so that when
we style the page, we can allow the user to know what page is they are
actively on.  One can do this by adding an IF statement.

IF statement:  Our IF statement here says if the variable $curpage is
equal to for example the“about” then  the php should echo active.
Declaring that page is active. The rest would be in CSS

The basename() function: returns the filename from the path. While the
Superglobal $_SERVER: holds information about headers, paths, and
script locations. Then the PHP_SELF returns the filename of the
currently executing script.

By combining the $_SERVER & PHP_SELF it executes the script.